cleanup TS done
