+++ vismop-vue/src/core/reactomeGraphs/reactomeOverviewNetwork/overviewTypes.d.ts
@@ -1,3 +1,5 @@
+import { baseNodeAttr, node, graphData } from '@/core/graphTypes';
+
 type fa2LayoutParams = {
   iterations: number;
   weightShared: number;
@@ -29,4 +31,82 @@ export type filterValues = {
   filterActive: boolean;
   inside: boolean;
   disable: boolean;
+};
+/**
+ * attributes of a node in the overview graph
+ */
+type overviewNodeAttr = baseNodeAttr & {
+  layoutX: number;
+  layoutY: number;
+  preFa2X: number;
+  preFa2Y: number;
+  yOnClusterFocus: number;
+  xOnClusterFocus: number;
+  clusterNum: number;
+  isRoot: boolean;
+  nodeType: 'root' | 'regular' | 'hierarchical' | 'cluster' | 'other';
+  nonHoverSize: number;
+  image: string;
+  imageHighDetail: string;
+  imageLowDetail: string;
+  hidden: boolean;
+  hierarchyHidden: boolean;
+  filterHidden: boolean;
+  zoomHidden: boolean;
+  clusterHidden: boolean;
+  clusterFixed: boolean;
+  forceLabel: boolean;
+  fcAverages: {
+    transcriptomics: number;
+    proteomics: number;
+    metabolomics: number;
+  };
+  nodeState: {
+    transcriptomics: { regulated: number; total: number };
+    proteomics: { regulated: number; total: number };
+    metabolomics: { regulated: number; total: number };
+  };
+  rootId: string;
+  parents: string[];
+  children: string[];
+  subtreeIds: string[];
+  visibleSubtree: boolean;
+};
+
+/**
+ * Defines an overview Node
+ */
+type overviewNode = node & {
+  attributes: overviewNodeAttr;
+};
+
+/**
+ * Type defining the points of the convex hull describing the clusters boundary
+ */
+type hullPoints = number[][][];
+
+/**
+ * Data needed for sigma to display the convex cluster hulls
+ */
+type additionalData = {
+  clusterAreas: {
+    hullPoints: hullPoints;
+    clusterColors: [number, number, number, number][];
+  };
+};
+
+/**
+ * Data decsribing the cluster hulls and their coloring
+ */
+type clusterData = {
+  normalHullPoints: hullPoints;
+  focusHullPoints: hullPoints;
+  clusterColors: [number, number, number, number][];
+};
+/**
+ * Defines the overview Graph Data
+ */
+type overviewGraphData = graphData & {
+  clusterData: clusterData;
+  nodes: overviewNode[];
 };
